[bigquery] use data_project and database_name correctly
