+++ vdsql/__init__.py
@@ -1,2 +1,3 @@
 from .__about__ import *
 from .ibis import *
+from .bigquery import *
